--- haalcentraal/build.gradle
@@ -1,84 +1,0 @@
-/*
- * Copyright 2015-2022 Ritense BV, the Netherlands.
- *
- * Licensed under EUPL, Version 1.2 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" basis,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-dockerCompose {
-    projectName = "haalcentraal"
-    isRequiredBy(integrationTesting)
-    isRequiredBy(securityTesting)
-    useComposeFiles.addAll("../docker-resources/docker-compose-base-test.yml", "docker-compose-override.yml")
-}
-
-dependencies {
-    implementation project(':core')
-    implementation project(':contract')
-    implementation project(':connector')
-
-    api "org.springframework.security:spring-security-core"
-    api "org.springframework.boot:spring-boot"
-    api "org.springframework.data:spring-data-commons"
-    api "javax.inject:javax.inject:1"
-
-    implementation "org.springframework.boot:spring-boot-starter"
-    implementation "org.springframework.boot:spring-boot-starter-web"
-    implementation "org.springframework.boot:spring-boot-starter-actuator"
-    implementation "org.springframework.boot:spring-boot-starter-data-jpa"
-    implementation "org.springframework.boot:spring-boot-starter-security"
-    implementation "org.springframework.boot:spring-boot-starter-validation"
-
-    implementation "com.fasterxml.jackson.module:jackson-module-afterburner"
-    implementation "com.fasterxml.jackson.module:jackson-module-parameter-names"
-    implementation "com.fasterxml.jackson.module:jackson-module-jaxb-annotations"
-    implementation "com.fasterxml.jackson.datatype:jackson-datatype-jsr310"
-
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
-    implementation "org.jetbrains.kotlin:kotlin-reflect"
-
-    implementation "io.github.microutils:kotlin-logging:2.0.4"
-
-    testImplementation "org.jetbrains.kotlin:kotlin-test"
-    testImplementation "org.junit.jupiter:junit-jupiter"
-    testImplementation "org.springframework.boot:spring-boot-starter-test"
-
-    implementation "org.springframework.boot:spring-boot-autoconfigure"
-
-    implementation "org.hibernate:hibernate-core"
-    implementation 'org.glassfish:javax.json:1.1.4'
-    implementation "org.springframework.boot:spring-boot-starter-webflux"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-reactive:1.6.0"
-
-    annotationProcessor "org.springframework.boot:spring-boot-autoconfigure-processor"
-
-    testImplementation project(path: ":core", configuration: 'testOutput')
-
-    testImplementation "org.springframework.boot:spring-boot-starter-test"
-    testImplementation "org.springframework.security:spring-security-test"
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.0"
-    testImplementation project(':test-utils-common')
-    testImplementation "com.squareup.okhttp3:mockwebserver:4.9.3"
-    testImplementation "com.squareup.okhttp3:okhttp:4.9.3"
-    testImplementation "com.nhaarman.mockitokotlin2:mockito-kotlin:2.2.0"
-    testImplementation "io.netty:netty-resolver-dns-native-macos:4.1.72.Final:osx-aarch_64"
-
-    jar {
-        enabled = true
-        manifest {
-            attributes("Implementation-Title": "Ritense Haalcentraal module")
-            attributes("Implementation-Version": projectVersion)
-        }
-    }
-}
-
-apply from: "gradle/publishing.gradle"
