+++ haalcentraal/haalcentraal-brp/.env
@@ -1,0 +1,1 @@
+COMPOSE_PROJECT_NAME=haalcentraal-brp
