+++ haalcentraal/haalcentraal-brp/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,4 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  com.ritense.haalcentraal.brp.autoconfigure.HaalCentraalAutoConfiguration, \
+  com.ritense.haalcentraal.brp.autoconfigure.HaalCentraalLiquibaseAutoConfiguration, \
+  com.ritense.haalcentraal.brp.autoconfigure.HaalCentraalSecurityAutoConfiguration
