+++ haalcentraal/haalcentraal-brp/src/test/kotlin/com/ritense/haalcentraal/brp/BaseTest.kt
@@ -1,0 +1,5 @@
+package com.ritense.haalcentraal.brp
+
+abstract class BaseTest {
+    fun readFileAsString(fileName: String): String = this::class.java.getResource(fileName)!!.readText(Charsets.UTF_8)
+}
