+++ haalcentraal/haalcentraal-brp/src/test/resources/config/application.yml
@@ -1,0 +1,44 @@
+spring:
+    datasource:
+        type: com.zaxxer.hikari.HikariDataSource
+        driver-class-name: com.mysql.cj.jdbc.Driver
+        url: jdbc:mysql://localhost:3319/haalcentraal-brp-test
+        username: valtimo
+        password: password
+        hikari:
+            auto-commit: false
+    liquibase:
+        enabled: false
+    jpa:
+        database-platform: org.hibernate.dialect.MySQL57Dialect
+        database: mysql
+        show_sql: true
+        open-in-view: false
+        properties:
+            hibernate:
+                hbm2ddl.auto: none
+                generate_statistics: false
+                naming-strategy: org.springframework.boot.orm.jpa.hibernate.SpringNamingStrategy
+                cache:
+                    use_second_level_cache: false
+                    use_query_cache: false
+                    region.factory_class: org.hibernate.cache.ehcache.SingletonEhCacheRegionFactory
+                format_sql: true
+                jdbc:
+                    time_zone: UTC
+                connection:
+                    provider_disables_autocommit: true
+        hibernate:
+            ddl-auto: none
+            naming:
+                physical-strategy: org.springframework.boot.orm.jpa.hibernate.SpringPhysicalNamingStrategy
+                implicit-strategy: org.springframework.boot.orm.jpa.hibernate.SpringImplicitNamingStrategy
+            use-new-id-generator-mappings: false
+
+spring-actuator:
+    username: test
+    password: test
+
+valtimo:
+    connector-encryption:
+        secret: "0123456789101112"
