--- haalcentraal/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.ritense.haalcentraal.autoconfigure.HaalCentraalAutoConfiguration, \
-  com.ritense.haalcentraal.autoconfigure.HaalCentraalSecurityAutoConfiguration
