--- haalcentraal/src/test/kotlin/com/ritense/haalcentraal/BaseTest.kt
@@ -1,5 +1,0 @@
-package com.ritense.haalcentraal
-
-abstract class BaseTest {
-    fun readFileAsString(fileName: String): String = this::class.java.getResource(fileName)!!.readText(Charsets.UTF_8)
-}
