+++ settings.gradle
@@ -33,6 +33,7 @@ include(
         "form-flow",
         "form-flow-valtimo",
         "form-link",
+        "haalcentraal:haalcentraal-brp",
         "document-generation",
         "document-generation:local-document-generation",
         "document-generation:smartdocuments",
